refactor(cli)!: split parsing and drop unsafe defaults
Break up the monolithic utils/parse.py so each subcommand owns its own
CLI arguments (COMMAND/HELP/add_arguments next to its run()), while the
globals and orchestration live in cli.py behind a single command
registry.

Remove the defaults for every value that decides where a run reads and
writes, when, and how it is billed. Previously these silently fell back
to a personal scratch dataset and fixed dates, so a misconfigured
production run could write to the wrong place. They are now required.
udfs_project is demoted from a global to a required argument on only the
two steps that use it.

Also move the now-example generate_incremental_fishing_events.sh into
examples/ (the run orchestrator drives fishing events through the python
pipe CLI, not this script) and update its references.

BREAKING CHANGE: --project, source/destination tables, start/end and
reference dates, labels, destination_table_prefix, nnet_score_night_loitering,
merged_table and udfs_project no longer have defaults and must be passed
explicitly.
