+++ pipe_events/constants.py
@@ -1,0 +1,5 @@
+from importlib.metadata import version
+
+PIPELINE_NAME = "pipe-events"
+PIPELINE_VERSION = version(PIPELINE_NAME)
+PIPELINE_DESCRIPTION = "Generate the incremental fishing events"
